+++ handler/CloseQuestionHandler.go
@@ -11,7 +11,7 @@ type CloseQuestionHandler struct {
 }
 
 func (self *CloseQuestionHandler) Get() {
-	qid, _ := self.GetInt(":qid")
+	qid, _ := self.GetInt64(":qid")
 	uid, _ := self.GetSession("userid").(int64)
 	role, _ := self.GetSession("userrole").(int64)
 	if qid > 0 && uid > 0 {
